+++ appmap/_implementation/__init__.py
@@ -1,5 +1,1 @@
-from . import configuration
-from .recording import recorder
-
-recorder.add_filter(configuration.BuiltinFilter.load_config())
-recorder.add_filter(configuration.ConfigFilter.load_config())
+from . import configuration  # noqa: F401
